+++ lomap/algorithms/value_iteration.py
@@ -53,7 +53,7 @@ def compute_mrp(p, backward=False):
 	else:
 		# Asynchronous backward value iteration
 		states_to_consider = set()
-		for s,_ in p.g.in_edges_iter(p.final):
+		for s,_ in p.g.in_edges(p.final):
 			states_to_consider.add(s)
 		states_to_consider -= p.final
 
@@ -91,11 +91,11 @@ def policy_synthesis(p, backward=False):
 	act_val = dict()
 	# Best actions for a state
 	act_max = dict()
-	for s in p.g.nodes_iter():
+	for s in p.g.nodes:
 		val[s] = 0
 		act_max[s] = p.available_controls(s)
 		act_val[s] = dict()
-	for s,_,d in p.g.out_edges_iter(data=True):
+	for s, _, d in p.g.out_edges(data=True):
 		act_val[s][d['control']] = 0
 
 	# Update val and act_val for final states
@@ -132,7 +132,7 @@ def policy_synthesis(p, backward=False):
 	else:
 		# Asynchronous backward value iteration
 		states_to_consider = set()
-		for s,_ in p.g.in_edges_iter(p.final):
+		for s,_ in p.g.in_edges(p.final):
 			states_to_consider.add(s)
 		states_to_consider -= p.final
 
@@ -170,7 +170,7 @@ def policy_synthesis(p, backward=False):
 	for s in p.final:
 		dist[s] = 0
 
-	edges_to_process = p.g.in_edges_iter(p.final, data=True)
+	edges_to_process = p.g.in_edges(p.final, data=True)
 	while edges_to_process:
 		new_edges_to_process = []
 		for s,t,d in edges_to_process: